requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
geojson==2.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.5
pandas==1.0.4
psycopg2-binary==2.8.5
pyaml==20.4.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
SQLAlchemy==1.3.17
Werkzeug==1.0.1
